Files for the project
For the utf2 file make sure that the usb is powered up and plugged into the computer, then as it is booted up make sure to drag the uf2 file and drop into the Raspberry Pi Pico. It would disconnect and then reboot on its own. Then you can continue with the rest of the project.